changed coordinate type from long to double

+++ 001 - Software Construction Fundamentals - GET/Chapter 009 - Factory For ADT/Point.java
@@ -1,33 +1,33 @@
 public class Point {
-    private long x = 0;
-    private long y = 0;
+    private double x = 0;
+    private double y = 0;
 
-    private final long ORIGIN_X = 0;
-    private final long ORIGIN_Y = 0;
+    private final double ORIGIN_X = 0;
+    private final double ORIGIN_Y = 0;
 
-    public Point(long x, long y){
+    public Point(double x, double y){
         this.x = x;
         this.y = y;
     }
 
-    public long getX(){
+    public double getX(){
         return x;
     }
 
-    public long getY(){
+    public double getY(){
         return y;
     }
     
     public double distanceFrom(Point p){
-        long xDifference = p.getX() - x;
-        long yDifference = p.getY() - y;
+        double xDifference = p.getX() - x;
+        double yDifference = p.getY() - y;
 
         double distance = Math.sqrt(Math.pow(xDifference, 2) + Math.pow(yDifference, 2));
         return distance;
     }
-    public double distanceFrom(long x, long y){
-        long xDifference = x - this.x;
-        long yDifference = y - this.y;
+    public double distanceFrom(double x, double y){
+        double xDifference = x - this.x;
+        double yDifference = y - this.y;
 
         double distance = Math.sqrt(Math.pow(xDifference, 2) + Math.pow(yDifference, 2));
         return distance;
